+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Updated
 1. Renamed _master_ branch to _main_ in line with current development practice.
+2. Replaced `nil` event pointer in `get-status` with zero value.
 
 
 ## [0.8.6](https://github.com/uhppoted/uhppoted-mqtt/releases/tag/v0.8.6) - 2023-08-30